chore: move jwt generator
Move the jwtGenerator into a separate package.
